+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/calendar/CalendarPanel.java
@@ -26,6 +26,7 @@ import javax.swing.SwingUtilities;
 import java.awt.BorderLayout;
 import java.awt.FlowLayout;
 import java.awt.GridLayout;
+import java.awt.Insets;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -166,9 +167,11 @@ public final class CalendarPanel extends JPanel {
   private final List<JLabel> dayFillLabels;
   private final JLabel formattedDateLabel;
   private final boolean includeTime;
+  private final boolean includeTodayButton;
 
   CalendarPanel(DefaultBuilder builder) {
     this.includeTime = builder.includeTime;
+    this.includeTodayButton = builder.includeTodayButton;
     LocalDateTime dateTime = builder.initialValue == null ? LocalDateTime.now() : builder.initialValue;
     yearValue = Value.value(dateTime.getYear(), dateTime.getYear());
     monthValue = Value.value(dateTime.getMonth(), dateTime.getMonth());
@@ -309,6 +312,12 @@ public final class CalendarPanel extends JPanel {
     Builder includeTime(boolean includeTime);
 
     /**
+     * @param includeTodayButton true if a 'Today' button for selecting the current date should be included
+     * @return this builder instance
+     */
+    Builder includeTodayButton(boolean includeTodayButton);
+
+    /**
      * @param keyboardShortcut the keyboard shortcut key
      * @param keyStroke the keyStroke to assign to the given shortcut key, null resets to the default one
      * @return this builder instance
@@ -327,6 +336,7 @@ public final class CalendarPanel extends JPanel {
 
     private LocalDateTime initialValue;
     private boolean includeTime = false;
+    private boolean includeTodayButton = false;
 
     @Override
     public Builder initialValue(LocalDate initialValue) {
@@ -343,6 +353,12 @@ public final class CalendarPanel extends JPanel {
     @Override
     public Builder includeTime(boolean includeTime) {
       this.includeTime = includeTime;
+      return this;
+    }
+
+    @Override
+    public Builder includeTodayButton(boolean includeTodayButton) {
+      this.includeTodayButton = includeTodayButton;
       return this;
     }
 
@@ -452,9 +468,12 @@ public final class CalendarPanel extends JPanel {
   }
 
   private Map<Integer, JToggleButton> createDayButtons() {
+    Insets margin = new Insets(0, 0, 0, 0);
+
     return dayStates.entrySet().stream()
             .collect(Collectors.toMap(Map.Entry::getKey, entry -> toggleButton(entry.getValue())
                     .text(Integer.toString(entry.getKey()))
+                    .margin(margin)
                     .build()));
   }
 
@@ -489,10 +508,13 @@ public final class CalendarPanel extends JPanel {
             .add(monthSpinner)
             .add(yearSpinner);
     if (includeTime) {
-      yearMonthHourMinutePanel.addAll(new JLabel(" "), hourSpinner, new JLabel(":", SwingConstants.CENTER), minuteSpinner);
+      yearMonthHourMinutePanel.addAll(hourSpinner, new JLabel(":", SwingConstants.CENTER), minuteSpinner);
+    }
+    if (includeTodayButton) {
+      yearMonthHourMinutePanel.add(createSelectTodayButton());
     }
 
-    return yearMonthHourMinutePanel.add(createSelectTodayButton()).build();
+    return yearMonthHourMinutePanel.build();
   }
 
   private JButton createSelectTodayButton() {
@@ -587,7 +609,7 @@ public final class CalendarPanel extends JPanel {
   }
 
   private void updateFormattedDate() {
-    formattedDateLabel.setText(dateFormatter.format(getLocalDateTime()) + (includeTime ? " " + timeFormatter.format(getLocalDateTime()) : ""));
+    formattedDateLabel.setText(dateFormatter.format(getLocalDateTime()) + (includeTime ? ", " + timeFormatter.format(getLocalDateTime()) : ""));
   }
 
   private void addKeyEvents(KeyboardShortcuts<KeyboardShortcut> keyboardShortcuts) {
@@ -723,7 +745,7 @@ public final class CalendarPanel extends JPanel {
 
   private static List<Item<Month>> createMonthItems() {
     return Arrays.stream(Month.values())
-            .map(month -> Item.item(month, month.getDisplayName(TextStyle.FULL, Locale.getDefault())))
+            .map(month -> Item.item(month, month.getDisplayName(TextStyle.SHORT, Locale.getDefault())))
             .collect(Collectors.toList());
   }